Add comment to clarify when getVideoObject works.
I've been playing with this all morning, and realized getVideoObject() seems to only work when I initialize vide via JS and not via html attributes. I'm guessing this is related to the earlier advice in the readme that js initialization is required to use mutation observers.

Thanks for the awesome library!